requirements.txt
torch
torchvision
numpy
Pillow
Flask
matplotlib
